barcampcanterbury.com2/pyproject.toml
[project]
name = "barcampcantebrury.com"
version = "0.0.0"
requires-python = ">= 3.14"
dependencies = [
  "mako",
  "pyyaml",  # PyYAML
]

[dependency-groups]
dev = [
  #"pytest",
  #"pytest-mypy",
  #"pytest-ruff",
]
